func/requirements.txt
azure-core==1.30.2
azure-functions==1.13.3
azure-storage-blob==12.21.0b1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
idna==3.8
isodate==0.6.1
pycparser==2.22
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
